requirements.txt
confluent-kafka
scapy>=2.4.5
pandas>=1.3.0
numpy
scikit-learn>=1.0.0
joblib>=1.2.0
kafka-python>=2.0.2
